separate adapter test runner with vitest
