+++ packages/adapter-drizzle/tests/mysql/index.test.ts
@@ -1,4 +1,4 @@
-import { runBasicTests } from "utils/adapter"
+import { runBasicTests } from "utils/adapter/vitest"
 import { DrizzleAdapter } from "../../src"
 import { db, sessions, verificationTokens, accounts, users } from "./schema"
 import { eq, and } from "drizzle-orm"
