+++ packages/adapter-drizzle/tests/sqlite/libsql.test.ts
@@ -1,4 +1,4 @@
-import { runBasicTests } from "utils/adapter"
+import { runBasicTests } from "utils/adapter/vitest"
 import { DrizzleAdapter } from "../../src"
 import {
   db,
